+++ src/strategies/base.ts
@@ -177,6 +177,22 @@ export abstract class BaseStrategy implements Strategy {
     });
   }
 
+  protected async buildPullRequestBody(
+    component: string | undefined,
+    newVersion: Version,
+    releaseNotesBody: string,
+    _conventionalCommits: ConventionalCommit[],
+    _latestRelease?: Release
+  ): Promise<PullRequestBody> {
+    return new PullRequestBody([
+      {
+        component,
+        version: newVersion,
+        notes: releaseNotesBody,
+      },
+    ]);
+  }
+
   /**
    * Builds a candidate release pull request
    * @param {Commit[]} commits Raw commits to consider for this release.
@@ -247,13 +263,13 @@ export abstract class BaseStrategy implements Strategy {
     const updatesWithExtras = mergeUpdates(
       updates.concat(...this.extraFileUpdates(newVersion))
     );
-    const pullRequestBody = new PullRequestBody([
-      {
-        component,
-        version: newVersion,
-        notes: releaseNotesBody,
-      },
-    ]);
+    const pullRequestBody = await this.buildPullRequestBody(
+      component,
+      newVersion,
+      releaseNotesBody,
+      conventionalCommits,
+      latestRelease
+    );
 
     return {
       title: pullRequestTitle,
@@ -325,6 +341,12 @@ export abstract class BaseStrategy implements Strategy {
     return new Map();
   }
 
+  protected async parsePullRequestBody(
+    pullRequestBody: string
+  ): Promise<PullRequestBody | undefined> {
+    return PullRequestBody.parse(pullRequestBody);
+  }
+
   /**
    * Given a merged pull request, build the candidate release.
    * @param {PullRequest} mergedPullRequest The merged release pull request.
@@ -360,7 +382,9 @@ export abstract class BaseStrategy implements Strategy {
       logger.error(`Bad branch name: ${mergedPullRequest.headBranchName}`);
       return;
     }
-    const pullRequestBody = PullRequestBody.parse(mergedPullRequest.body);
+    const pullRequestBody = await this.parsePullRequestBody(
+      mergedPullRequest.body
+    );
     if (!pullRequestBody) {
       logger.error('Could not parse pull request body as a release PR');
       return;